requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pyttsx3>=2.90
torch>=2.0.0
torchaudio>=2.0.0
sounddevice>=0.4.6
librosa>=0.10.0
transformers>=4.30.0
accelerate>=0.20.0
PyQt6>=6.5.0
